PN-27 all permission checks updated

+++ apps/backend/src/app/product/product.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { FindOptionsRelations } from 'typeorm';
 
 import { Product, User } from '@rentigo/models';
 import { CreateProductDto, UpdateProductDto } from '@rentigo/dto';
 
+import { Operations } from '@rentigo/constants';
 import { AddressService } from '../address';
 import { RentingPolicyService } from '../renting-policy';
 import { ResourceService } from '../resource';
@@ -15,7 +16,6 @@ import { ProductRepository } from './product.repository';
 export class ProductService {
 	constructor(
 		private readonly productRepository: ProductRepository,
-		private readonly userRepository: UserRepository,
 		private readonly addressService: AddressService,
 		private readonly rentingPolicyService: RentingPolicyService,
 		private readonly resourceService: ResourceService,
@@ -35,8 +35,11 @@ export class ProductService {
 		return this.productRepository.findOne({ where: { id }, relations });
 	}
 
-	async update(updateProductDto: UpdateProductDto): Promise<Product> {
+	async update(user: User, updateProductDto: UpdateProductDto): Promise<Product> {
 		let product = await this.findOne(updateProductDto.id);
+		if (product.lender.id !== user.id) {
+			throw new ForbiddenException();
+		}
 		product.imageUrls.map((imageUrl) => this.resourceService.remove(imageUrl.id));
 		const updatedProduct = await this.dtoToEntity(updateProductDto);
 		updatedProduct.id = product.id;
@@ -44,10 +47,24 @@ export class ProductService {
 		return product;
 	}
 
-	async remove(id: string): Promise<Product> {
-		const product = await this.findOne(id);
-		product.lender = null;
-		return this.productRepository.save(product);
+	async remove(user: User, id: string): Promise<Product> {
+		const isPermitted = await this.hasPermissionTo(Operations.DELETE, user, id);
+		if (!isPermitted) {
+			throw new ForbiddenException();
+		}
+		return this.productRepository.softRemoveOneBy({ id });
+	}
+
+	async hasPermissionTo(_: Operations, user: User, id: string) {
+		try {
+			const userWithProduct = await this.productRepository.findOne({
+				relations: { lender: true },
+				where: { id, lender: { id: user.id } },
+			});
+			return !!userWithProduct;
+		} catch (error) {
+			return false;
+		}
 	}
 
 	private async dtoToEntity(dto: CreateProductDto): Promise<Product> {
